+++ packages/apps/package.json
@@ -18,5 +18,8 @@
     "@polkadot/ui-signer": "^0.24.32",
     "@types/react-tooltip": "^3.9.1",
     "react-tooltip": "^3.9.2"
+  },
+  "devDependencies": {
+    "@hot-loader/react-dom": "^16.8.3"
   }
 }
