+++ packages/apps/src/SideBar/index.tsx
@@ -21,25 +21,10 @@ import getLogo from './logos';
 
 type Props = I18nProps;
 
-type State = {
-  isCollapsed: boolean
-};
-
-class SideBar extends React.PureComponent<Props, State> {
-  state: State;
-
-  constructor (props: Props) {
-    super(props);
-
-    const state = store.get('sidebar') || {};
-    this.state = {
-      isCollapsed: false,
-      ...state
-    };
-  }
+class SideBar extends React.PureComponent<Props> {
 
   render () {
-    const { isCollapsed } = this.state;
+    const { isCollapsed } = this.props;
 
     return (
       <div className={`apps--SideBar ${isCollapsed ? 'collapsed' : 'expanded'}`}>
@@ -65,16 +50,8 @@ class SideBar extends React.PureComponent<Props, State> {
     );
   }
 
-  private collapse = (): void => {
-    this.setState(({ isCollapsed }: State) => ({
-      isCollapsed: !isCollapsed
-    }), () => {
-      store.set('sidebar', this.state);
-    });
-  }
-
   private renderCollapse () {
-    const { isCollapsed } = this.state;
+    const { isCollapsed } = this.props;
 
     return (
       <div className='apps--SideBar-collapse'>
@@ -82,14 +59,14 @@ class SideBar extends React.PureComponent<Props, State> {
           icon={`angle double ${isCollapsed ? 'right' : 'left'}`}
           isBasic
           isCircular
-          onClick={this.collapse}
+          onClick={this.props.collapse}
         />
       </div>
     );
   }
 
   private renderLogo () {
-    const { isCollapsed } = this.state;
+    const { isCollapsed } = this.props;
     const logo = getLogo(isCollapsed);
 
     return (
@@ -102,7 +79,7 @@ class SideBar extends React.PureComponent<Props, State> {
   }
 
   private renderRoutes () {
-    const { isCollapsed } = this.state;
+    const { isCollapsed } = this.props;
     const { t } = this.props;
 
     return routing.routes.map((route, index) => (
@@ -141,7 +118,7 @@ class SideBar extends React.PureComponent<Props, State> {
     return (
       <div
         className='apps--SideBar-toggle'
-        onClick={this.collapse}
+        onClick={this.props.collapse}
       >
       </div>
     );
